Matias1

+++ src/main/java/com/PoloDeSalud/UBB/rest/NoticiaController.java
@@ -4,7 +4,6 @@ import com.PoloDeSalud.UBB.model.Noticia;
 import com.PoloDeSalud.UBB.service.NoticiaService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-
 import java.util.List;
 
 @RestController
@@ -30,7 +29,7 @@ public class NoticiaController {
     }
 
     @DeleteMapping("/{id}")
-    public void eliminarNoticia(@PathVariable Long id) {
+    public void eliminarNoticia(@PathVariable int id) {
         noticiaService.eliminar(id);
     }
 }
